requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.42.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
